#1 Split config delegator from higher level service
