+++ pkg/plugin/dogu/config/delegator.go
@@ -2,6 +2,7 @@ package config
 
 import (
 	"fmt"
+
 	"github.com/cloudogu/kubectl-ces-plugin/pkg/logger"
 
 	"github.com/phayes/freeport"
@@ -22,14 +23,13 @@ type doguConfigurationDelegator struct {
 	editor    doguConfigurationEditor
 }
 
-func newDoguConfigurationDelegator(doguName string, namespace string, restConfig *rest.Config) (*doguConfigurationDelegator, error) {
+func newDelegator(doguName string, namespace string, restConfig *rest.Config) (*doguConfigurationDelegator, error) {
 	freePort, err := freeport.GetFreePort()
 	if err != nil {
 		return nil, err
 	}
 
-	var forward portForwarder
-	forward = portforward.NewPortForwarder(restConfig, portforward.ServiceType, types.NamespacedName{Namespace: namespace, Name: "etcd"}, freePort, 4001)
+	forward := portforward.NewPortForwarder(restConfig, portforward.ServiceType, types.NamespacedName{Namespace: namespace, Name: "etcd"}, freePort, 4001)
 
 	endpoint := fmt.Sprintf("http://localhost:%d", freePort)
 	etcd, err := registry.New(core.Registry{
@@ -72,6 +72,7 @@ func newDoguConfigurationDelegator(doguName string, namespace string, restConfig
 	}, nil
 }
 
+// Delegate executes the given dogu configuration call against a previously configured registry.
 func (dcd *doguConfigurationDelegator) Delegate(doguConfigCall func(dogu *core.Dogu, editor doguConfigurationEditor) error) error {
 	err := dcd.forwarder.ExecuteWithPortForward(func() error {
 
@@ -97,41 +98,4 @@ func (dcd *doguConfigurationDelegator) Delegate(doguConfigCall func(dogu *core.D
 	}
 
 	return nil
-}
-
-func NewDoguConfigService(doguName, namespace string, restConfig *rest.Config) (*doguConfigService, error) {
-	delegator, err := newDoguConfigurationDelegator(doguName, namespace, restConfig)
-	if err != nil {
-		return nil, err
-	}
-
-	return &doguConfigService{
-		delegator: delegator,
-	}, nil
-}
-
-type doguConfigService struct {
-	delegator *doguConfigurationDelegator
-}
-
-func (s doguConfigService) EditAllInteractive() error {
-	return s.delegator.Delegate(func(dogu *core.Dogu, editor doguConfigurationEditor) error {
-		return editor.EditConfiguration(dogu.Configuration)
-	})
-}
-
-func (s doguConfigService) Edit(registryKey string, registryValue string) error {
-	panic("todo")
-}
-
-func (s doguConfigService) Delete(registryKey string) error {
-	panic("todo")
-}
-
-func (s doguConfigService) GetAllForDogu() (map[string]string, error) {
-	panic("todo")
-}
-
-func (s doguConfigService) GetValue(registryKey string) (string, error) {
-	panic("todo")
 }
